requirements.txt
streamlit
streamlit-aggrid
numpy
matplotlib
pandasai
streamlit-elements
streamlit-option-menu
streamlit-pandas-profiling
plotly
pymongo
tk
